[Refactoring] Move AddEquatableConformance to its own file
