+++ lib/Refactoring/Utils.h
@@ -20,5 +20,9 @@ StringRef correctNameInternal(ASTContext &Ctx, StringRef Name,
                               ArrayRef<ValueDecl *> AllVisibles);
 
 StringRef correctNewDeclName(DeclContext *DC, StringRef Name);
+
+/// If \p NTD is a protocol, return all the protocols it inherits from. If it's
+/// a type, return all the protocols it conforms to.
+SmallVector<ProtocolDecl *, 2> getAllProtocols(NominalTypeDecl *NTD);
 }
 } // namespace swift
